+++ crates/tui/src/util.rs
@@ -9,6 +9,7 @@ use crossterm::event;
 use editor_command::EditorBuilder;
 use futures::{FutureExt, future};
 use mime::Mime;
+use notify::{RecommendedWatcher, RecursiveMode, Watcher, event::ModifyKind};
 use slumber_core::{template::Prompt, util::doc_link};
 use slumber_util::{ResultTraced, paths::expand_home};
 use std::{
@@ -148,7 +149,7 @@ pub fn spawn(future: impl 'static + Future<Output = ()>) -> JoinHandle<()> {
             _ = future => {
                 // Assume the task updated _something_ visible to the user,
                 // so trigger a redraw here
-                ViewContext::messages_tx().send(Message::Tick);
+                ViewContext::messages_tx().send(Message::Draw);
             },
             _ = CANCEL_TOKEN.cancelled() => {},
         }
@@ -380,6 +381,50 @@ pub async fn confirm(
     messages_tx.send(Message::ConfirmStart(confirm));
     // Error means we got ghosted :( RUDE!
     rx.await.unwrap_or_default()
+}
+
+/// Spawn a file system watcher that watches a given set of files. When any file
+/// in the set is modified, call the given callback. Return the spawned watcher.
+/// **The watcher will stop when it is dropped, so hang onto the return
+/// value!!**
+pub fn watch_files(
+    paths: &[&Path],
+    on_change: impl Fn(notify::Event) + 'static + Send + Sync,
+) -> notify::Result<RecommendedWatcher> {
+    let on_file_event = move |result: notify::Result<notify::Event>| {
+        match result {
+            // Only reload if the file *content* changes
+            Ok(
+                event @ notify::Event {
+                    kind: notify::EventKind::Modify(ModifyKind::Data(_)),
+                    ..
+                },
+            ) => on_change(event),
+            // Do nothing for other event kinds
+            Ok(_) => {}
+            Err(err) => {
+                error!(error = %err, "Error watching file");
+            }
+        }
+    };
+
+    // Spawn a watcher for all paths in the source tree
+    notify::recommended_watcher(on_file_event)
+        .map(|mut watcher| {
+            for path in paths {
+                // If any single file fails, just ignore it
+                watcher
+                    .watch(path, RecursiveMode::NonRecursive)
+                    .with_context(|| format!("Error watching file {path:?}"))
+                    .traced()
+                    .ok();
+            }
+            info!(paths = ?paths, ?watcher, "Watching file(s) for changes");
+            watcher
+        })
+        .inspect_err(|err| {
+            error!(error = %err, "Error starting watcher for file(s)");
+        })
 }
 
 #[cfg(test)]